Fixed namespaces; moved files to corresponding folders
