+++ TRPZ_Labs/TRPZ_Labs/BusinessLayer/Services/Interfaces/IShopService.cs
@@ -1,7 +1,7 @@
 ﻿using OrderingGoods.BusinessLayer.DomainModels;
 using System.Collections.Generic;
 
-namespace OrderingGoods.BusinessLayer
+namespace OrderingGoods.BusinessLayer.Services
 {
     public interface IShopService
     {
