+++ TRPZ_Labs/TRPZ_Labs/PresentationLayer/ApplicationViewModel.cs
@@ -8,6 +8,7 @@ using System.Text;
 using System.Threading.Tasks;
 using OrderingGoods.BusinessLayer;
 using OrderingGoods.BusinessLayer.DomainModels;
+using OrderingGoods.BusinessLayer.Services;
 
 namespace OrderingGoods.PresentationLayer
 {
