+++ TRPZ_Labs/TRPZ_Labs/PresentationLayer/OrderingGoodsWindow.xaml.cs
