+++ TRPZ_Labs/TRPZ_Labs/PresentationLayer/RelayCommand.cs
